Cast strlen to int to avoid compiler warning

+++ apps/s_client.c
@@ -2172,7 +2172,7 @@ int s_client_main(int argc, char **argv)
              * HTTP/d.d ddd Reason text\r\n
              */
             mbuf_len = BIO_gets(fbio, mbuf, BUFSIZZ);
-            if (mbuf_len < strlen("HTTP/1.0 200")) {
+            if (mbuf_len < (int)strlen("HTTP/1.0 200")) {
                 BIO_printf(bio_err,
                            "%s: HTTP CONNECT failed, insufficient response "
                            "from proxy (got %d octets)\n", prog, mbuf_len);
